distinction "mainClass" et "defaultClass" dans les factory
