+++ project/view/site.php
@@ -1,9 +1,9 @@
 <?php
 namespace project\view;
 
-use base\view\View;
+use base\view\HtmlView;
 
-class site extends View
+class site extends HtmlView
 {
 	public function __construct()
     {
